==> code/4d. SSP Effect Projections Adaptation - Income.Rmd <==
---
title: "SSP Projection"
subtitle: "Income Adaptation"
author: "Moritz Schwarz and Felix Pretis"
version: "October 2020"
output:
  html_document:
    df_print: paged
    toc: true
    toc_depth: 2
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = here::here("output")) })
---

```{r,purl = FALSE}
knitr::opts_chunk$set(message = F,echo = T,fig.width = 10)
```


```{r "setup", message=FALSE}
library(tidyverse)
library(data.table)
library(gets)
library(here)
library(vroom)
library(MASS)

library(extrafont)
library(RColorBrewer)
library(Cairo)
library(viridis)


rm(list = ls())
select <- dplyr::select
```


**To create Server Script: Save first then**
```{r,eval= TRUE, purl = FALSE}
knitr::purl(here("code","4d. SSP Effect Projections Adaptation - Income.Rmd"),output = here("code","4d. Server_SSP_Adaptation.R"))
```

# SSP
## Process 
```{r}
SSP_wide <- vroom(file=here("data","raw","SspDb_country_data_2013-06-12.csv"))

SSP_wide %>% 
  filter(MODEL == "OECD Env-Growth") %>% 
  filter(VARIABLE %in% c("Population","GDP|PPP")) %>% 
  pivot_longer(cols = -c(MODEL,SCENARIO,REGION,VARIABLE,UNIT),names_to = "YEAR") %>% 
  mutate(YEAR = YEAR %>% as.character %>% as.integer) %>% 
  pivot_wider(id_cols = c(MODEL,SCENARIO,REGION,YEAR),names_from = VARIABLE) %>% 
  
  mutate(GDP = `GDP|PPP`*1000000000,
         `GDP|PPP`=NULL,
         POPULATION = `Population`*1000000,
         `Population`=NULL,
         GDP_pc = GDP/POPULATION,
         MODEL = NULL,
         SCENARIO = gsub("_v9_130325","",SCENARIO)) %>% 
  
  filter(!YEAR == 2000) %>% 
  drop_na -> ssp
```

## Interpolate SSPs using a simple spline
```{r}
ssp %>% 
  select(REGION) %>% 
  unique %>% 
  slice(rep(1:n(), each = 90)) %>%
  group_by(REGION) %>% 
  mutate(YEAR = 1:90) %>% 
  ungroup() %>% 
  mutate(YEAR = YEAR + 2009) %>% 
  full_join(ssp %>% 
              select(-GDP,-POPULATION) %>% 
              filter(YEAR >2005) %>% 
              pivot_wider(id_cols = c(REGION,YEAR),values_from = "GDP_pc",names_from = "SCENARIO"),by=c("REGION","YEAR")) %>% 
  group_by(REGION) %>% 
  mutate(SSP1 = zoo::na.spline(SSP1,YEAR),
         SSP2 = zoo::na.spline(SSP2,YEAR),
         SSP3 = zoo::na.spline(SSP3,YEAR),
         SSP4 = zoo::na.spline(SSP4,YEAR),
         SSP5 = zoo::na.spline(SSP5,YEAR)) %>% 
  ungroup %>% 
  rename(iso = REGION,
         year = YEAR) -> ssp

#write.csv(ssp,"ssp/ssp_interpolated.csv",row.names = F)
```

```{r}
ssp %>% 
  group_by(iso) %>% 
  mutate(across(.cols = starts_with("SSP"),.fns = ~./lag(.))) %>% 
  ungroup %>% 
  drop_na %>% 
  rename_with(.cols = starts_with("SSP"),.fn = ~paste0(.,"_g"))-> ssp_growth
```

Prepare the SSP dataframe so that it has: 1) the SSP growth rates 2) the SSP Climate Placeholders starting in 2011
```{r}
ssp_growth %>% 
  #left_join(ssp,by=c("iso","year")) %>% 
  left_join(ssp %>% 
              filter(year == 2011) %>% 
              rename_with(.cols = starts_with("SSP"),.fn = ~paste0(.,"_clim")),by=c("iso","year")) -> ssp_ready
```


```{r}
coefsamples <- 100
```

# Adaptation GETS

## Load Coefficients

### Standard (for restriction)
```{r}
load(here("data","temp","standard_gets_model.RData"))
####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_standard <- MASS::mvrnorm(n = coefsamples, mu = standard_model %>% coef(), Sigma = standard_model %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

# If we are just using one coefsample, we simply use the mean estimate from the model
if(coefsamples==1){
  standard_model %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_standard
}
```


### Log Interaction All
```{r}
load(here("data","temp","log_interaction_model_all.RData"))

####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_interaction_all_log <- MASS::mvrnorm(n = coefsamples, mu = interaction_model_all_log %>% coef(), Sigma = interaction_model_all_log %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

if(coefsamples==1){
  interaction_model_all_log %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_interaction_all_log
}
```


### Log Interaction Selected
```{r}
load(here("data","temp","log_interaction_model_selected.RData"))

####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_interaction_selected_log <- MASS::mvrnorm(n = coefsamples, mu = interaction_model_selected_log %>% coef(), Sigma = interaction_model_selected_log %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

if(coefsamples==1){
  interaction_model_selected_log %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_interaction_selected_log
}
```


## Collect Relevant Information

```{r}
selection_coefs_standard %>% 
  names %>% 
  gsub("_coef","",.) %>% 
  unique -> climate_vars

selection_coefs_standard %>% 
  names %>% 
  gsub("_coef|_2","",.) %>% 
  unique -> climate_vars_lin

selection_coefs_standard %>% 
  names %>% 
  grep("_2",.,value = TRUE) %>% 
  gsub("_coef|_2","",.) %>% 
  unique -> climate_vars_sq

selection_coefs_interaction_all_log %>% 
  names %>% 
  gsub("_coef","",.) %>% 
  grep("_int",.,value = TRUE) %>% 
  gsub("_int","",.) %>% 
  unique -> interaction_vars_all

selection_coefs_interaction_selected_log %>%
  names %>%
  gsub("_coef","",.) %>%
  grep("_int",.,value = TRUE) %>%
  gsub("_int","",.) %>%
  unique -> interaction_vars_selected
```

## Load climate data
```{r}
climate <- vroom(here("data","temp","corrected_anomaly_climatedata.csv"))
climate %>% 
  select(model,rcp,ensemble,final_temp,iso,year,all_of(climate_vars_lin),
         all_of(paste0(climate_vars_sq,"_2"))) %>% 
  drop_na -> climate_subset
#mutate(across(.cols = climate_vars_sq,.fns = list(`2` = ~.^2))) -> climate_subset
rm(climate)
```

# Unrestriced

**October 2020**
Not used anymore

<!-- ## Project (Server) -->
<!-- ```{r, purl=F, eval=F} -->
<!-- library(doMC) -->
<!-- registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1 -->
<!-- foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% { -->
<!--   #for (v in 1:coefsamples){ -->
<!--   print(v) -->

<!--   effect_interaction_selected_log <- climate_subset -->
<!--   # Calculate the climate effect -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   for(var in interaction_vars_selected){ -->
<!--     effect_interaction_selected_log %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_int_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_int_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   # Calculate the standard climate effect -->
<!--   effect_standard <- climate_subset -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_standard %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_standard,.) -> effect_standard -->
<!--   } -->

<!--   effect_standard %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std -->

<!--   effect_interaction_selected_log %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>%  -->
<!--     bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int -->

<!--   done <- ssp_ready %>%  -->
<!--     filter(year==2011) %>%  -->
<!--     left_join(effect_int %>%  -->
<!--                 filter(year == 2011) %>%  -->
<!--                 select(-model,-ensemble),by = c("iso","year")) %>%  -->

<!--     #drop_na %>%  -->
<!--     mutate(realisation = v) %>%  -->
<!--     relocate(c(realisation,final_temp), .after = year) -->

<!--   for(i in 2012:2099){ -->
<!--     print(paste0("Realisation ",v," Year: ",i)) -->
<!--     #i=2018 -->

<!--     done %>%  -->
<!--       filter(year == i-1) %>%  -->
<!--       select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running) -->

<!--       inner_join(ssp_ready %>%  -->
<!--                    select(-ends_with("_clim")) %>%  -->
<!--                    filter(year == i-1), by = c("iso","year")) %>%  -->

<!--       inner_join(effect_int %>%  -->
<!--                    filter(year == i-1) %>%  -->
<!--                    mutate(realisation = v) %>%  -->
<!--                    select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>%  -->

<!--       # Aggregate the growth rates:  -->
<!--       # stdeffect is the normal effect from the model without interaction -->
<!--       # logeffect is the generic identifier for an effect in the interaction projection -->
<!--       # effects that need to be interacted with GDP carry additional "_int_" in front of "logeffect" -->
<!--       mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))), -->

<!--              total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))), -->
<!--              total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>%  -->

<!--       # Calculate the interaction growth rate (climate coefficients in the interaction model + climate*GDP coefficients) -->
<!--       mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect*log(SSP1_clim), -->
<!--              interaction_effect_SSP2 = total_climlogeffect + total_inteffect*log(SSP2_clim), -->
<!--              interaction_effect_SSP3 = total_climlogeffect + total_inteffect*log(SSP3_clim), -->
<!--              interaction_effect_SSP4 = total_climlogeffect + total_inteffect*log(SSP4_clim), -->
<!--              interaction_effect_SSP5 = total_climlogeffect + total_inteffect*log(SSP5_clim)) %>%  -->

<!--       # Checking if the standard growth rate (without interaction) is higher. The higher one is chosen -->
<!--       mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1), -->
<!--              total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2), -->
<!--              total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3), -->
<!--              total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4), -->
<!--              total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>% -->

<!--       # Growth process: current GDP * growth rate for this year -->
<!--       mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1), -->
<!--              SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2), -->
<!--              SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3), -->
<!--              SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4), -->
<!--              SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>% -->

<!--       mutate(year = i) %>%  -->

<!--       #select(-starts_with("total_")) %>%  -->

<!--       bind_rows(done,.) -> done -->

<!--   } -->

<!--   save(done,file = here("data","temp","projections", -->
<!--                         paste0("full_SSP_interaction_log_selected_unrestricted_",v,".RData"))) -->
<!--   rm(done,i) -->
<!-- } -->
<!-- ``` -->



# Restriction Max Overall GDP

**October 2020**
Not used anymore

<!-- ```{r} -->
<!-- ssp %>%  -->
<!--   filter(year==2011) %>%  -->
<!--   summarise(across(starts_with("SSP"),max)) -> max_overall_gdp -->
<!-- ``` -->

<!-- ## Project (Server) -->
<!-- ```{r} -->
<!-- library(doMC) -->
<!-- registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1 -->
<!-- foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% { -->
<!--   #for (v in 1:coefsamples){ -->
<!--   print(v) -->

<!--   effect_interaction_selected_log <- climate_subset -->
<!--   # Calculate the climate effect -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   for(var in interaction_vars_selected){ -->
<!--     effect_interaction_selected_log %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_int_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_int_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   # Calculate the standard climate effect -->
<!--   effect_standard <- climate_subset -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_standard %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_standard,.) -> effect_standard -->
<!--   } -->

<!--   effect_standard %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std -->

<!--   effect_interaction_selected_log %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>%  -->
<!--     bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int -->

<!--   done <- ssp_ready %>%  -->
<!--     filter(year==2011) %>%  -->
<!--     left_join(effect_int %>%  -->
<!--                 filter(year == 2011) %>%  -->
<!--                 select(-model,-ensemble),by = c("iso","year")) %>%  -->

<!--     #drop_na %>%  -->
<!--     mutate(realisation = v) %>%  -->
<!--     relocate(c(realisation,final_temp), .after = year) -->

<!--   for(i in 2012:2099){ -->
<!--     print(paste0("Realisation ",v," Year: ",i)) -->
<!--     #i=2018 -->

<!--     done %>%  -->
<!--       filter(year == i-1) %>%  -->
<!--       select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running) -->

<!--       inner_join(ssp_ready %>%  -->
<!--                    select(-ends_with("_clim")) %>%  -->
<!--                    filter(year == i-1), by = c("iso","year")) %>%  -->

<!--       left_join(effect_int %>%  -->
<!--                   filter(year == i-1) %>%  -->
<!--                   mutate(realisation = v) %>%  -->
<!--                   select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>%  -->


<!--       mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))), -->

<!--              total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))), -->
<!--              total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>%  -->

<!--       mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(ifelse(SSP1_clim > max_overall_gdp$SSP1,max_overall_gdp$SSP1,SSP1_clim)), -->
<!--              interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(ifelse(SSP2_clim > max_overall_gdp$SSP2,max_overall_gdp$SSP2,SSP2_clim)), -->
<!--              interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(ifelse(SSP3_clim > max_overall_gdp$SSP3,max_overall_gdp$SSP3,SSP3_clim)), -->
<!--              interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(ifelse(SSP4_clim > max_overall_gdp$SSP4,max_overall_gdp$SSP4,SSP4_clim)), -->
<!--              interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(ifelse(SSP5_clim > max_overall_gdp$SSP5,max_overall_gdp$SSP5,SSP5_clim))) %>%  -->

<!--       mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1), -->
<!--              total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2), -->
<!--              total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3), -->
<!--              total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4), -->
<!--              total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>% -->

<!--       mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1), -->
<!--              SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2), -->
<!--              SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3), -->
<!--              SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4), -->
<!--              SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>% -->

<!--       mutate(year = i) %>%  -->


<!--       bind_rows(done,.) -> done -->

<!--   } -->

<!--   save(done,file = here("data","temp","projections", -->
<!--                         paste0("full_SSP_interaction_log_selected_restricted_",v,".RData"))) -->
<!--   rm(done,i) -->
<!-- } -->
<!-- ``` -->




# Alternative Restriction 1: Baseline Restriction and Max GDP

Idea: Because we are only attenuating the effect of climate change, when the standard effect would suggest that GDP decreases, the Adaptation equivalent can only increase by a maximum of the baseline. 

Option 1: if country would lose, best could do is baseline. 
Combined: can do no worse than the no-adaptation baseline, but if the signs don't agree can't do better than baseline growth

```{r}
ssp %>% 
  filter(year==2011) %>% 
  summarise(across(starts_with("SSP"),max)) -> max_overall_gdp
```

## Project (Server)
```{r}
library(doMC)
registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1
foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% {
  #for (v in 1:coefsamples){
  print(v)
  
  effect_interaction_selected_log <- climate_subset
  # Calculate the climate effect
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  for(var in interaction_vars_selected){
    effect_interaction_selected_log %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_int_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_int_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  # Calculate the standard climate effect
  effect_standard <- climate_subset
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_standard %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_standard,.) -> effect_standard
  }
  
  effect_standard %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std
  
  effect_interaction_selected_log %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>% 
    bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int
  
  done <- ssp_ready %>% 
    filter(year==2011) %>% 
    left_join(effect_int %>% 
                filter(year == 2011) %>% 
                select(-model,-ensemble),by = c("iso","year")) %>% 
    
    #drop_na %>% 
    mutate(realisation = v) %>% 
    relocate(c(realisation,final_temp), .after = year)
  
  for(i in 2012:2099){
    print(paste0("Realisation ",v," Year: ",i))
    #i=2018
    
    done %>% 
      filter(year == i-1) %>% 
      select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running)
      
      inner_join(ssp_ready %>% 
                   select(-ends_with("_clim")) %>% 
                   filter(year == i-1), by = c("iso","year")) %>% 
      
      left_join(effect_int %>% 
                  filter(year == i-1) %>% 
                  mutate(realisation = v) %>% 
                  select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>% 
      
      
      mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))),
             
             total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))),
             total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>% 
      
      mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(ifelse(SSP1_clim > max_overall_gdp$SSP1,max_overall_gdp$SSP1,SSP1_clim)),
             interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(ifelse(SSP2_clim > max_overall_gdp$SSP2,max_overall_gdp$SSP2,SSP2_clim)),
             interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(ifelse(SSP3_clim > max_overall_gdp$SSP3,max_overall_gdp$SSP3,SSP3_clim)),
             interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(ifelse(SSP4_clim > max_overall_gdp$SSP4,max_overall_gdp$SSP4,SSP4_clim)),
             interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(ifelse(SSP5_clim > max_overall_gdp$SSP5,max_overall_gdp$SSP5,SSP5_clim))) %>% 
      
      mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1),
             total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2),
             total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3),
             total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4),
             total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>%
      
      # mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1),
      #        SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2),
      #        SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3),
      #        SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4),
      #        SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>%
      
      
      
      # in words: if there is a sign difference, take take the smaller of the interaction or the base effect
      mutate(
        SSP1_clim = SSP1_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP1)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP1_g<(total_effect_this_year_SSP1+1),SSP1_g,(total_effect_this_year_SSP1+1)),
          no = SSP1_g + total_effect_this_year_SSP1
        ),
        SSP2_clim = SSP2_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP2)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP2_g<(total_effect_this_year_SSP2+1),SSP2_g,(total_effect_this_year_SSP2+1)),
          no = SSP2_g + total_effect_this_year_SSP2
        ),
        SSP3_clim = SSP3_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP3)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP3_g<(total_effect_this_year_SSP3+1),SSP3_g,(total_effect_this_year_SSP3+1)),
          no = SSP3_g + total_effect_this_year_SSP3
        ),
        SSP4_clim = SSP4_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP4)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP4_g<(total_effect_this_year_SSP4+1),SSP4_g,(total_effect_this_year_SSP4+1)),
          no = SSP4_g + total_effect_this_year_SSP4
        ),
        SSP5_clim = SSP5_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP5)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP5_g<(total_effect_this_year_SSP5+1),SSP5_g,(total_effect_this_year_SSP5+1)),
          no = SSP5_g + total_effect_this_year_SSP5
        )
      ) %>% 
      
      mutate(year = i) %>% 
      
      
      bind_rows(done,.) -> done
    
  }
  
 
  save(done,file = here("data","temp","projections",
                        paste0("full_SSP_interaction_log_selected_altrestr1_",v,".RData")))
  rm(done,i)
}
```



# Alternative Restriction 2: Baseline Restriction

Idea: Because we are only attenuating the effect of climate change, when the standard effect would suggest that GDP decreases, the Adaptation equivalent can only increase by a maximum of the baseline. 

Option 1: if country would lose, best could do is baseline. 
Combined: can do no worse than the no-adaptation baseline, but if the signs don't agree can't do better than baseline growth

## Project (Server)
```{r}
library(doMC)
registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1
foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% {
  #for (v in 1:coefsamples){
  print(v)
  
  effect_interaction_selected_log <- climate_subset
  # Calculate the climate effect
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  for(var in interaction_vars_selected){
    effect_interaction_selected_log %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_int_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_int_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  # Calculate the standard climate effect
  effect_standard <- climate_subset
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_standard %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_standard,.) -> effect_standard
  }
  
  effect_standard %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std
  
  effect_interaction_selected_log %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>% 
    bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int
  
  done <- ssp_ready %>% 
    filter(year==2011) %>% 
    left_join(effect_int %>% 
                filter(year == 2011) %>% 
                select(-model,-ensemble),by = c("iso","year")) %>% 
    
    #drop_na %>% 
    mutate(realisation = v) %>% 
    relocate(c(realisation,final_temp), .after = year)
  
  for(i in 2012:2099){
    print(paste0("Realisation ",v," Year: ",i))
    #i=2018
    
    done %>% 
      filter(year == i-1) %>% 
      select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running)
      
      inner_join(ssp_ready %>% 
                   select(-ends_with("_clim")) %>% 
                   filter(year == i-1), by = c("iso","year")) %>% 
      
      left_join(effect_int %>% 
                  filter(year == i-1) %>% 
                  mutate(realisation = v) %>% 
                  select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>% 
      
      
      mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))),
             
             total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))),
             total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>% 
      
      mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(SSP1_clim),
             interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(SSP2_clim),
             interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(SSP3_clim),
             interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(SSP4_clim),
             interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(SSP5_clim)) %>% 
      
      mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1),
             total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2),
             total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3),
             total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4),
             total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>%
      
      # mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1),
      #        SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2),
      #        SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3),
      #        SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4),
      #        SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>%
      
      
      
      # in words: if there is a sign difference, take take the smaller of the interaction or the base effect
      mutate(
        SSP1_clim = SSP1_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP1)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP1_g<(total_effect_this_year_SSP1+1),SSP1_g,(total_effect_this_year_SSP1+1)),
          no = SSP1_g + total_effect_this_year_SSP1
        ),
        SSP2_clim = SSP2_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP2)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP2_g<(total_effect_this_year_SSP2+1),SSP2_g,(total_effect_this_year_SSP2+1)),
          no = SSP2_g + total_effect_this_year_SSP2
        ),
        SSP3_clim = SSP3_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP3)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP3_g<(total_effect_this_year_SSP3+1),SSP3_g,(total_effect_this_year_SSP3+1)),
          no = SSP3_g + total_effect_this_year_SSP3
        ),
        SSP4_clim = SSP4_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP4)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP4_g<(total_effect_this_year_SSP4+1),SSP4_g,(total_effect_this_year_SSP4+1)),
          no = SSP4_g + total_effect_this_year_SSP4
        ),
        SSP5_clim = SSP5_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP5)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP5_g<(total_effect_this_year_SSP5+1),SSP5_g,(total_effect_this_year_SSP5+1)),
          no = SSP5_g + total_effect_this_year_SSP5
        )
      ) %>% 
      
      mutate(year = i) %>% 
      
      
      bind_rows(done,.) -> done
    
  }
  
  save(done,file = here("data","temp","projections",
                        paste0("full_SSP_interaction_log_selected_altrestr2_",v,".RData")))
  rm(done,i)
}
```

# Adaptation LASSO

## Load Coefficients

### Standard (for restriction)
```{r}
load(here("data","temp","standard_lasso_model.RData"))
####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_standard <- MASS::mvrnorm(n = coefsamples, mu = lasso_standard_model %>% coef(), Sigma = lasso_standard_model %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

# If we are just using one coefsample, we simply use the mean estimate from the model
if(coefsamples==1){
  lasso_standard_model %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_standard
}
```


### Log Interaction All
```{r}
load(here("data","temp","lasso_log_interaction_model_all.RData"))

####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_interaction_all_log <- MASS::mvrnorm(n = coefsamples, mu = lasso_interaction_model_all_log %>% coef(), Sigma = lasso_interaction_model_all_log %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

if(coefsamples==1){
  lasso_interaction_model_all_log %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_interaction_all_log
}
```


### Log Interaction Selected
```{r}
load(here("data","temp","lasso_log_interaction_model_selected.RData"))

####simulate coefficients, draw them form joint normal distribution
set.seed(123)
selection_coefs_interaction_selected_log <- MASS::mvrnorm(n = coefsamples, mu = lasso_interaction_model_selected_log %>% coef(), Sigma = lasso_interaction_model_selected_log %>% vcov()) %>%
  {if (is.vector(.)) t(.) else .} %>%
  data.frame() %>%
  rename_all(~ tolower(.)) %>%
  select(-l1.diff.ln_gdp_cap, -starts_with(c("year", "iis", "time", "iso"))) %>%
  rename_all(~ paste0(., "_coef"))

if(coefsamples==1){
  lasso_interaction_model_selected_log %>%
    coef %>%
    data.frame(variable = names(.),
               coefficient = .,
               row.names = NULL) %>%
    pivot_wider(names_from = "variable",values_from = "coefficient") %>%
    rename_all(~tolower(.)) %>%
    select(-l1.diff.ln_gdp_cap,-starts_with(c("year","iis","time","iso"))) %>% 
    rename_all(~paste0(.,"_coef")) -> selection_coefs_interaction_selected_log
}
```


## Collect Relevant Information

```{r}
selection_coefs_standard %>% 
  names %>% 
  gsub("_coef","",.) %>% 
  unique -> climate_vars

selection_coefs_standard %>% 
  names %>% 
  gsub("_coef|_2","",.) %>% 
  unique -> climate_vars_lin

selection_coefs_standard %>% 
  names %>% 
  grep("_2",.,value = TRUE) %>% 
  gsub("_coef|_2","",.) %>% 
  unique -> climate_vars_sq

selection_coefs_interaction_all_log %>% 
  names %>% 
  gsub("_coef","",.) %>% 
  grep("_int",.,value = TRUE) %>% 
  gsub("_int","",.) %>% 
  unique -> interaction_vars_all

selection_coefs_interaction_selected_log %>%
  names %>%
  gsub("_coef","",.) %>%
  grep("_int",.,value = TRUE) %>%
  gsub("_int","",.) %>%
  unique -> interaction_vars_selected
```

## Load climate data
```{r}
climate <- vroom(here("data","temp","corrected_anomaly_climatedata.csv"))
climate %>% 
  select(model,rcp,ensemble,final_temp,iso,year,all_of(climate_vars_lin),
         all_of(paste0(climate_vars_sq,"_2"))) %>% 
  drop_na -> climate_subset
#mutate(across(.cols = climate_vars_sq,.fns = list(`2` = ~.^2))) -> climate_subset
rm(climate)
```

# Unrestriced

**October 2020**
Not used anymore


<!-- ## Project (Server) -->
<!-- ```{r} -->
<!-- library(doMC) -->
<!-- registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1 -->
<!-- foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% { -->
<!--   #for (v in 1:coefsamples){ -->
<!--   print(v) -->

<!--   effect_interaction_selected_log <- climate_subset -->
<!--   # Calculate the climate effect -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   for(var in interaction_vars_selected){ -->
<!--     effect_interaction_selected_log %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_int_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_int_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   # Calculate the standard climate effect -->
<!--   effect_standard <- climate_subset -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_standard %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_standard,.) -> effect_standard -->
<!--   } -->

<!--   effect_standard %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std -->

<!--   effect_interaction_selected_log %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>%  -->
<!--     bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int -->

<!--   done <- ssp_ready %>%  -->
<!--     filter(year==2011) %>%  -->
<!--     left_join(effect_int %>%  -->
<!--                 filter(year == 2011) %>%  -->
<!--                 select(-model,-ensemble),by = c("iso","year")) %>%  -->

<!--     #drop_na %>%  -->
<!--     mutate(realisation = v) %>%  -->
<!--     relocate(c(realisation,final_temp), .after = year) -->

<!--   for(i in 2012:2099){ -->
<!--     print(paste0("Realisation ",v," Year: ",i)) -->
<!--     #i=2018 -->

<!--     done %>%  -->
<!--       filter(year == i-1) %>%  -->
<!--       select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running) -->

<!--       inner_join(ssp_ready %>%  -->
<!--                    select(-ends_with("_clim")) %>%  -->
<!--                    filter(year == i-1), by = c("iso","year")) %>%  -->

<!--       inner_join(effect_int %>%  -->
<!--                    filter(year == i-1) %>%  -->
<!--                    mutate(realisation = v) %>%  -->
<!--                    select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>%  -->

<!--       # Aggregate the growth rates:  -->
<!--       # stdeffect is the normal effect from the model without interaction -->
<!--       # logeffect is the generic identifier for an effect in the interaction projection -->
<!--       # effects that need to be interacted with GDP carry additional "_int_" in front of "logeffect" -->
<!--       mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))), -->

<!--              total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))), -->
<!--              total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>%  -->

<!--       # Calculate the interaction growth rate (climate coefficients in the interaction model + climate*GDP coefficients) -->
<!--       mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect*log(SSP1_clim), -->
<!--              interaction_effect_SSP2 = total_climlogeffect + total_inteffect*log(SSP2_clim), -->
<!--              interaction_effect_SSP3 = total_climlogeffect + total_inteffect*log(SSP3_clim), -->
<!--              interaction_effect_SSP4 = total_climlogeffect + total_inteffect*log(SSP4_clim), -->
<!--              interaction_effect_SSP5 = total_climlogeffect + total_inteffect*log(SSP5_clim)) %>%  -->

<!--       # Checking if the standard growth rate (without interaction) is higher. The higher one is chosen -->
<!--       mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1), -->
<!--              total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2), -->
<!--              total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3), -->
<!--              total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4), -->
<!--              total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>% -->

<!--       # Growth process: current GDP * growth rate for this year -->
<!--       mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1), -->
<!--              SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2), -->
<!--              SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3), -->
<!--              SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4), -->
<!--              SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>% -->

<!--       mutate(year = i) %>%  -->

<!--       #select(-starts_with("total_")) %>%  -->

<!--       bind_rows(done,.) -> done -->

<!--   } -->


<!--   save(done,file = here("data","temp","projections", -->
<!--                         paste0("full_SSP_lasso_interaction_log_selected_unrestricted_",v,".RData"))) -->
<!--   rm(done,i) -->
<!-- } -->
<!-- ``` -->


# Restriction Max Overall GDP

**October 2020**
Not used anymore



<!-- ```{r} -->
<!-- ssp %>%  -->
<!--   filter(year==2011) %>%  -->
<!--   summarise(across(starts_with("SSP"),max)) -> max_overall_gdp -->
<!-- ``` -->

<!-- ## Project (Server) -->
<!-- ```{r} -->
<!-- library(doMC) -->
<!-- registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1 -->
<!-- foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% { -->
<!--   #for (v in 1:coefsamples){ -->
<!--   print(v) -->

<!--   effect_interaction_selected_log <- climate_subset -->
<!--   # Calculate the climate effect -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   for(var in interaction_vars_selected){ -->
<!--     effect_interaction_selected_log %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_interaction_selected_log %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_int_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_int_effect"))  %>%  -->
<!--       bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log -->
<!--   } -->

<!--   # Calculate the standard climate effect -->
<!--   effect_standard <- climate_subset -->
<!--   for(var in climate_vars){ -->
<!--     #print(var) -->
<!--     climate_subset %>%  -->
<!--       select(all_of(var)) %>%  -->
<!--       pull %>% "*"(selection_coefs_standard %>%  -->
<!--                      slice(v) %>%  -->
<!--                      select(all_of(paste0(var,"_coef"))) %>%  -->
<!--                      pull) %>% -->
<!--       as_tibble %>%  -->
<!--       rename_all(~paste0(var,"_effect"))  %>%  -->
<!--       bind_cols(effect_standard,.) -> effect_standard -->
<!--   } -->

<!--   effect_standard %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std -->

<!--   effect_interaction_selected_log %>%  -->
<!--     rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>%  -->
<!--     bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int -->

<!--   done <- ssp_ready %>%  -->
<!--     filter(year==2011) %>%  -->
<!--     left_join(effect_int %>%  -->
<!--                 filter(year == 2011) %>%  -->
<!--                 select(-model,-ensemble),by = c("iso","year")) %>%  -->

<!--     #drop_na %>%  -->
<!--     mutate(realisation = v) %>%  -->
<!--     relocate(c(realisation,final_temp), .after = year) -->

<!--   for(i in 2012:2099){ -->
<!--     print(paste0("Realisation ",v," Year: ",i)) -->
<!--     #i=2018 -->

<!--     done %>%  -->
<!--       filter(year == i-1) %>%  -->
<!--       select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running) -->

<!--       inner_join(ssp_ready %>%  -->
<!--                    select(-ends_with("_clim")) %>%  -->
<!--                    filter(year == i-1), by = c("iso","year")) %>%  -->

<!--       left_join(effect_int %>%  -->
<!--                   filter(year == i-1) %>%  -->
<!--                   mutate(realisation = v) %>%  -->
<!--                   select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>%  -->


<!--       mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))), -->

<!--              total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))), -->
<!--              total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>%  -->

<!--       mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(ifelse(SSP1_clim > max_overall_gdp$SSP1,max_overall_gdp$SSP1,SSP1_clim)), -->
<!--              interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(ifelse(SSP2_clim > max_overall_gdp$SSP2,max_overall_gdp$SSP2,SSP2_clim)), -->
<!--              interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(ifelse(SSP3_clim > max_overall_gdp$SSP3,max_overall_gdp$SSP3,SSP3_clim)), -->
<!--              interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(ifelse(SSP4_clim > max_overall_gdp$SSP4,max_overall_gdp$SSP4,SSP4_clim)), -->
<!--              interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(ifelse(SSP5_clim > max_overall_gdp$SSP5,max_overall_gdp$SSP5,SSP5_clim))) %>%  -->

<!--       mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1), -->
<!--              total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2), -->
<!--              total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3), -->
<!--              total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4), -->
<!--              total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>% -->

<!--       mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1), -->
<!--              SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2), -->
<!--              SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3), -->
<!--              SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4), -->
<!--              SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>% -->

<!--       mutate(year = i) %>%  -->


<!--       bind_rows(done,.) -> done -->

<!--   } -->


<!--   save(done,file = here("data","temp","projections", -->
<!--                         paste0("full_SSP_lasso_interaction_log_selected_restricted_",v,".RData"))) -->
<!--   rm(done,i) -->
<!-- } -->
<!-- ``` -->



# Alternative Restriction 1: Baseline Restriction and Max GDP

Idea: Because we are only attenuating the effect of climate change, when the standard effect would suggest that GDP decreases, the Adaptation equivalent can only increase by a maximum of the baseline. 

Option 1: if country would lose, best could do is baseline. 
Combined: can do no worse than the no-adaptation baseline, but if the signs don't agree can't do better than baseline growth

```{r}
ssp %>% 
  filter(year==2011) %>% 
  summarise(across(starts_with("SSP"),max)) -> max_overall_gdp
```

## Project (Server)
```{r}
library(doMC)
registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1
foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% {
  #for (v in 1:coefsamples){
  print(v)
  
  effect_interaction_selected_log <- climate_subset
  # Calculate the climate effect
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  for(var in interaction_vars_selected){
    effect_interaction_selected_log %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_int_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_int_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  # Calculate the standard climate effect
  effect_standard <- climate_subset
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_standard %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_standard,.) -> effect_standard
  }
  
  effect_standard %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std
  
  effect_interaction_selected_log %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>% 
    bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int
  
  done <- ssp_ready %>% 
    filter(year==2011) %>% 
    left_join(effect_int %>% 
                filter(year == 2011) %>% 
                select(-model,-ensemble),by = c("iso","year")) %>% 
    
    #drop_na %>% 
    mutate(realisation = v) %>% 
    relocate(c(realisation,final_temp), .after = year)
  
  for(i in 2012:2099){
    print(paste0("Realisation ",v," Year: ",i))
    #i=2018
    
    done %>% 
      filter(year == i-1) %>% 
      select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running)
      
      inner_join(ssp_ready %>% 
                   select(-ends_with("_clim")) %>% 
                   filter(year == i-1), by = c("iso","year")) %>% 
      
      left_join(effect_int %>% 
                  filter(year == i-1) %>% 
                  mutate(realisation = v) %>% 
                  select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>% 
      
      
      mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))),
             
             total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))),
             total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>% 
      
      mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(ifelse(SSP1_clim > max_overall_gdp$SSP1,max_overall_gdp$SSP1,SSP1_clim)),
             interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(ifelse(SSP2_clim > max_overall_gdp$SSP2,max_overall_gdp$SSP2,SSP2_clim)),
             interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(ifelse(SSP3_clim > max_overall_gdp$SSP3,max_overall_gdp$SSP3,SSP3_clim)),
             interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(ifelse(SSP4_clim > max_overall_gdp$SSP4,max_overall_gdp$SSP4,SSP4_clim)),
             interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(ifelse(SSP5_clim > max_overall_gdp$SSP5,max_overall_gdp$SSP5,SSP5_clim))) %>% 
      
      mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1),
             total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2),
             total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3),
             total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4),
             total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>%
      
      # mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1),
      #        SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2),
      #        SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3),
      #        SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4),
      #        SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>%
      
      
      
      # in words: if there is a sign difference, take take the smaller of the interaction or the base effect
      mutate(
        SSP1_clim = SSP1_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP1)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP1_g<(total_effect_this_year_SSP1+1),SSP1_g,(total_effect_this_year_SSP1+1)),
          no = SSP1_g + total_effect_this_year_SSP1
        ),
        SSP2_clim = SSP2_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP2)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP2_g<(total_effect_this_year_SSP2+1),SSP2_g,(total_effect_this_year_SSP2+1)),
          no = SSP2_g + total_effect_this_year_SSP2
        ),
        SSP3_clim = SSP3_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP3)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP3_g<(total_effect_this_year_SSP3+1),SSP3_g,(total_effect_this_year_SSP3+1)),
          no = SSP3_g + total_effect_this_year_SSP3
        ),
        SSP4_clim = SSP4_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP4)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP4_g<(total_effect_this_year_SSP4+1),SSP4_g,(total_effect_this_year_SSP4+1)),
          no = SSP4_g + total_effect_this_year_SSP4
        ),
        SSP5_clim = SSP5_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP5)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP5_g<(total_effect_this_year_SSP5+1),SSP5_g,(total_effect_this_year_SSP5+1)),
          no = SSP5_g + total_effect_this_year_SSP5
        )
      ) %>% 
      
      mutate(year = i) %>% 
      
      
      bind_rows(done,.) -> done
    
  }
  

  save(done,file = here("data","temp","projections",
                        paste0("full_SSP_lasso_interaction_log_selected_altrestr1_",v,".RData")))
  rm(done,i)
}
```



# Alternative Restriction 2: Baseline Restriction

Idea: Because we are only attenuating the effect of climate change, when the standard effect would suggest that GDP decreases, the Adaptation equivalent can only increase by a maximum of the baseline. 

Option 1: if country would lose, best could do is baseline. 
Combined: can do no worse than the no-adaptation baseline, but if the signs don't agree can't do better than baseline growth

## Project (Server)
```{r}
library(doMC)
registerDoMC(if(coefsamples < detectCores()){coefsamples} else {detectCores()-1})  # coefsamples if enough cores available - otherwise total-1
foreach(v=1:coefsamples,.packages = loadedNamespaces()) %dopar% {
  #for (v in 1:coefsamples){
  print(v)
  
  effect_interaction_selected_log <- climate_subset
  # Calculate the climate effect
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  for(var in interaction_vars_selected){
    effect_interaction_selected_log %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_interaction_selected_log %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_int_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_int_effect"))  %>% 
      bind_cols(effect_interaction_selected_log,.) -> effect_interaction_selected_log
  }
  
  # Calculate the standard climate effect
  effect_standard <- climate_subset
  for(var in climate_vars){
    #print(var)
    climate_subset %>% 
      select(all_of(var)) %>% 
      pull %>% "*"(selection_coefs_standard %>% 
                     slice(v) %>% 
                     select(all_of(paste0(var,"_coef"))) %>% 
                     pull) %>%
      as_tibble %>% 
      rename_all(~paste0(var,"_effect"))  %>% 
      bind_cols(effect_standard,.) -> effect_standard
  }
  
  effect_standard %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","stdeffect",.)) -> effect_std
  
  effect_interaction_selected_log %>% 
    rename_with(.cols = contains("effect"),.fn = ~gsub("effect","logeffect",.)) %>% 
    bind_cols(effect_std %>% select(contains("stdeffect"))) -> effect_int
  
  done <- ssp_ready %>% 
    filter(year==2011) %>% 
    left_join(effect_int %>% 
                filter(year == 2011) %>% 
                select(-model,-ensemble),by = c("iso","year")) %>% 
    
    #drop_na %>% 
    mutate(realisation = v) %>% 
    relocate(c(realisation,final_temp), .after = year)
  
  for(i in 2012:2099){
    print(paste0("Realisation ",v," Year: ",i))
    #i=2018
    
    done %>% 
      filter(year == i-1) %>% 
      select(iso, year, realisation, final_temp,ends_with("_clim")) %>% #contains(SSP_scenarios_running)
      
      inner_join(ssp_ready %>% 
                   select(-ends_with("_clim")) %>% 
                   filter(year == i-1), by = c("iso","year")) %>% 
      
      left_join(effect_int %>% 
                  filter(year == i-1) %>% 
                  mutate(realisation = v) %>% 
                  select(-model,-ensemble,-rcp),by = c("iso","year","final_temp","realisation")) %>% 
      
      
      mutate(total_stdeffect = rowSums(select(.,contains("stdeffect"))),
             
             total_climlogeffect = rowSums(select(.,contains("logeffect"),-contains("_int_"))),
             total_inteffect =  rowSums(select(.,contains("_int_logeffect")))) %>% 
      
      mutate(interaction_effect_SSP1 = total_climlogeffect + total_inteffect * log(SSP1_clim),
             interaction_effect_SSP2 = total_climlogeffect + total_inteffect * log(SSP2_clim),
             interaction_effect_SSP3 = total_climlogeffect + total_inteffect * log(SSP3_clim),
             interaction_effect_SSP4 = total_climlogeffect + total_inteffect * log(SSP4_clim),
             interaction_effect_SSP5 = total_climlogeffect + total_inteffect * log(SSP5_clim)) %>% 
      
      mutate(total_effect_this_year_SSP1 = ifelse(total_stdeffect > interaction_effect_SSP1,total_stdeffect,interaction_effect_SSP1),
             total_effect_this_year_SSP2 = ifelse(total_stdeffect > interaction_effect_SSP2,total_stdeffect,interaction_effect_SSP2),
             total_effect_this_year_SSP3 = ifelse(total_stdeffect > interaction_effect_SSP3,total_stdeffect,interaction_effect_SSP3),
             total_effect_this_year_SSP4 = ifelse(total_stdeffect > interaction_effect_SSP4,total_stdeffect,interaction_effect_SSP4),
             total_effect_this_year_SSP5 = ifelse(total_stdeffect > interaction_effect_SSP5,total_stdeffect,interaction_effect_SSP5)) %>%
      
      # mutate(SSP1_clim = SSP1_clim*(SSP1_g + total_effect_this_year_SSP1),
      #        SSP2_clim = SSP2_clim*(SSP2_g + total_effect_this_year_SSP2),
      #        SSP3_clim = SSP3_clim*(SSP3_g + total_effect_this_year_SSP3),
      #        SSP4_clim = SSP4_clim*(SSP4_g + total_effect_this_year_SSP4),
      #        SSP5_clim = SSP5_clim*(SSP5_g + total_effect_this_year_SSP5)) %>%
      
      
      
      # in words: if there is a sign difference, take take the smaller of the interaction or the base effect
      mutate(
        SSP1_clim = SSP1_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP1)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP1_g<(total_effect_this_year_SSP1+1),SSP1_g,(total_effect_this_year_SSP1+1)),
          no = SSP1_g + total_effect_this_year_SSP1
        ),
        SSP2_clim = SSP2_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP2)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP2_g<(total_effect_this_year_SSP2+1),SSP2_g,(total_effect_this_year_SSP2+1)),
          no = SSP2_g + total_effect_this_year_SSP2
        ),
        SSP3_clim = SSP3_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP3)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP3_g<(total_effect_this_year_SSP3+1),SSP3_g,(total_effect_this_year_SSP3+1)),
          no = SSP3_g + total_effect_this_year_SSP3
        ),
        SSP4_clim = SSP4_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP4)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP4_g<(total_effect_this_year_SSP4+1),SSP4_g,(total_effect_this_year_SSP4+1)),
          no = SSP4_g + total_effect_this_year_SSP4
        ),
        SSP5_clim = SSP5_clim* ifelse(
          test = sign(total_stdeffect)==-1 & sign(interaction_effect_SSP5)==1, # if sign between stdeffect != interaction 
          yes = ifelse(SSP5_g<(total_effect_this_year_SSP5+1),SSP5_g,(total_effect_this_year_SSP5+1)),
          no = SSP5_g + total_effect_this_year_SSP5
        )
      ) %>% 
      
      mutate(year = i) %>% 
      
      
      bind_rows(done,.) -> done
    
  }
  
  save(done,file = here("data","temp","projections",
                        paste0("full_SSP_lasso_interaction_log_selected_altrestr2_",v,".RData")))
  rm(done,i)
}
```
